MH first go at tidying up Aidan's C1 elements
-- moved duplicated (and essentially identical) RotatedBoundaryHelper
   class (defined in FvK and KS libraries) to (soon to be renamed; see
   below) my_geom_object.h

-- deleted FvK and KS's upgrade_element_to_curved(...) member function; not
   needed; just calls the underlying version in CurvableBellElement
   anyway.

-- changed CurvilineGeomObject to C1CurviLine, a shallow wrapper to
   the curviline class used for the triangle mesh generation (both have
   to represent exactly the same boundary anyway). the only new
   function is get_zeta(...) which I've implemented with a (virtual)
   Newton method; can be overloaded.

-- changed tolerance for connectivity check of nodes on boundaries to 1e-10.

-- my_geom_object.h: File kept alive (really needs to be renamed to
   something like C1_helper.h) and implemented CurvilineGeomObject,
   constructed from triangle mesh equivalent. Also parked RotatedBoundaryHelper
   there. Deleted all the rest since none of these objects need to be
   constructed separately; they're already contained in the
   GeomObjects that define the curvilinear boundaries of the triangle mesh.

-- Created TemplateFreeCurvableBellElement with pure virtual
   definitions of upgrade_element_to_curved(...) and
   rotated_boundary_helper_pt(). Bolted underneath the CurvableBellElement.
