+++ src/c1_foeppl_von_karman/foeppl_von_karman_curvable_bell_elements.h
@@ -46,350 +46,352 @@
 
 namespace oomph
 {
-  //===start of rotation helper class=========================================
-  /// Helper class to contain all the rotation information in the element.
-  class RotatedBoundaryHelper
-  {
-  public:
-    /// Constructor: just initialise the member data to their defaults (zeros)
-    RotatedBoundaryHelper(FiniteElement* const& parent_element_pt)
-      : Parent_element_pt(parent_element_pt),
-        Nnode(Parent_element_pt->nvertex_node()),
-        Boundary_coordinate_of_node(3, 0.0),
-        Nodal_boundary_parametrisation_pt(3, 0),
-        Rotation_matrix_at_node(3, DenseMatrix<double>(6, 6, 0.0))
-    {
-    }
+ 
+  // //===start of rotation helper class=========================================
+  // /// Helper class to contain all the rotation information in the element.
+  // class RotatedBoundaryHelper
+  // {
+  // public:
+  //   /// Constructor: just initialise the member data to their defaults (zeros)
+  //   RotatedBoundaryHelper(FiniteElement* const& parent_element_pt)
+  //     : Parent_element_pt(parent_element_pt),
+  //       Nnode(Parent_element_pt->nvertex_node()),
+  //       Boundary_coordinate_of_node(3, 0.0),
+  //       Nodal_boundary_parametrisation_pt(3, 0),
+  //       Rotation_matrix_at_node(3, DenseMatrix<double>(6, 6, 0.0))
+  //   {
+  //   }
 
-    /// Destructor
-    ~RotatedBoundaryHelper() {}
+  //   /// Destructor
+  //   ~RotatedBoundaryHelper() {}
 
-    CurvilineGeomObject* nodal_boundary_parametrisation_pt(
-      const unsigned& j_node)
-    {
-      return Nodal_boundary_parametrisation_pt[j_node];
-    }
+  //   C1CurviLine* nodal_boundary_parametrisation_pt(
+  //     const unsigned& j_node)
+  //   {
+  //     return Nodal_boundary_parametrisation_pt[j_node];
+  //   }
 
+  //  // hierher isn't this the same as in fkv?
 
-    /// Add a new boundary parametrisation to nodes all the nodes in the
-    /// vector node_on_boundary
-    void set_nodal_boundary_parametrisation(
-      const Vector<unsigned>& node_on_boundary,
-      const Vector<double>& boundary_coord_of_node,
-      CurvilineGeomObject* const& boundary_parametrisation_pt)
-    {
-      // Loop over all the nodes in node_on_boundary and add the boundary
-      // pointer to their vector of boundaries
-      unsigned n_node = node_on_boundary.size();
-      for (unsigned j = 0; j < n_node; j++)
-      {
-        // The j-th node on the boundary
-        unsigned j_node = node_on_boundary[j];
+  //   /// Add a new boundary parametrisation to nodes all the nodes in the
+  //   /// vector node_on_boundary
+  //   void set_nodal_boundary_parametrisation(
+  //     const Vector<unsigned>& node_on_boundary,
+  //     const Vector<double>& boundary_coord_of_node,
+  //     C1CurviLine* const& boundary_parametrisation_pt)
+  //   {
+  //     // Loop over all the nodes in node_on_boundary and add the boundary
+  //     // pointer to their vector of boundaries
+  //     unsigned n_node = node_on_boundary.size();
+  //     for (unsigned j = 0; j < n_node; j++)
+  //     {
+  //       // The j-th node on the boundary
+  //       unsigned j_node = node_on_boundary[j];
 
-        // Set the boundary parametrisation data pointer for this node
-        Nodal_boundary_parametrisation_pt[j_node] = boundary_parametrisation_pt;
+  //       // Set the boundary parametrisation data pointer for this node
+  //       Nodal_boundary_parametrisation_pt[j_node] = boundary_parametrisation_pt;
 
-        // Set the coordinate of node j on this boundary
-        Boundary_coordinate_of_node[j_node] = boundary_coord_of_node[j];
+  //       // Set the coordinate of node j on this boundary
+  //       Boundary_coordinate_of_node[j_node] = boundary_coord_of_node[j];
 
-        update_rotation_matrices();
-      } // end of loop over nodes in node_on_boundary [j]
-    } // end of set_nodal_boundary_parametrisation()
-
-
-    /// Update all rotation matrices (checks if they are needed unless flag is
-    /// true)
-    void update_rotation_matrices()
-    {
-      // [zdec] hard coded the three vertex nodes
-      unsigned n_vertex = 3;
-      // Loop over each vertex
-      for (unsigned j_node = 0; j_node < n_vertex; j_node++)
-      {
-        // If this node does not have a parametrisation (the pointer is still
-        // null) skip over it, otherwise we go on to fill out the rotation
-        // matrix
-        if (!nodal_boundary_parametrisation_pt(j_node))
-        {
-          continue;
-        }
-
-        // Initialise the two basis vectors and their jacobians
-        Vector<Vector<double>> bi(2, Vector<double>(2, 0.0));
-        Vector<DenseMatrix<double>> dbidx(2, DenseMatrix<double>(2, 2, 0.0));
-
-        // Our new coordinate system:
-        //     (l, s)=(normal component, tangent component)
-        // which we define in terms of basis vectors (rescaled)
-        //     ni=dxi/dl / |n|           <-- Jacobian col 1
-        //     ti=dxi/ds / |t|           <-- Jacobian col 2
-        // and their derivatives
-        //     dnidxj=d/dxj(dxi/dl / |n|) <-- Hessian `col' 1
-        //     dtidxj=d/dxj(dxi/ds / |t|) <-- Hessian `col' 2
-
-        // [zdec] we use i and j for brevity
-        // but it should be alpha & beta
-        // Need to write up how the transformation is done
-
-        // Storage for our basis and derivatives
-        Vector<double> ni(2, 0.0);
-        Vector<double> ti(2, 0.0);
-        Vector<double> dnids(2, 0.0);
-        Vector<double> dtids(2, 0.0);
-
-        // All tensors assumed evaluated on the boundary
-        // Jacobian of inverse mapping
-        DenseMatrix<double> jac_inv(2, 2, 0.0);
-        // Hessian of mapping [zdec] (not needed because...)
-        Vector<DenseMatrix<double>> hess(2, DenseMatrix<double>(2, 2, 0.0));
-        // Hessian of inverse mapping [zdec] (...this can be found by
-        // hand)
-        Vector<DenseMatrix<double>> hess_inv(2, DenseMatrix<double>(2, 2, 0.0));
-
-        // The basis is defined in terms of the boundary parametrisation
-        Vector<double> boundary_coord = {Boundary_coordinate_of_node[j_node]};
-        CurvilineGeomObject* boundary_pt =
-          Nodal_boundary_parametrisation_pt[j_node];
-        Vector<double> x(2, 0.0);
-        Vector<double> dxids(2, 0.0);
-        Vector<double> d2xids2(2, 0.0);
-
-        // Get position (debug)
-        boundary_pt->position(boundary_coord, x);
-        // Get tangent vector
-        boundary_pt->dposition(boundary_coord, dxids);
-        // Get second derivative
-        boundary_pt->d2position(boundary_coord, d2xids2);
-
-        double mag_t = sqrt(dxids[0] * dxids[0] + dxids[1] * dxids[1]);
-        // ti is the normalised tangent vector
-        ti[0] = dxids[0] / mag_t;
-        ti[1] = dxids[1] / mag_t;
-        // Derivative of (normalised) tangent
-        dtids[0] = d2xids2[0] / std::pow(mag_t, 2) -
-                   (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[0] /
-                     std::pow(mag_t, 4);
-        dtids[1] = d2xids2[1] / std::pow(mag_t, 2) -
-                   (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[1] /
-                     std::pow(mag_t, 4);
-        // n = (t x e_z) implies
-        ni[0] = ti[1];
-        ni[1] = -ti[0];
-        // Same for dnids
-        dnids[0] = dtids[1];
-        dnids[1] = -dtids[0];
-
-        // Need inverse of mapping to calculate ds/dxi ----------------
-        //   /  dx/dl  dx/ds  \ -1  ___  __1__ /  dy/ds -dx/ds \ .
-        //   \  dy/dl  dy/ds  /     ---   det  \ -dy/dl  dx/dl /
-        //
-        //                          ___  /  dl/dx  dl/dy  \ .
-        //                          ---  \  ds/dx  ds/dy  /
-        //
-        // Fill out inverse of Jacobian
-        double det = (ni[0] * ti[1] - ni[1] * ti[0]);
-        jac_inv(0, 0) = ti[1] / det;
-        jac_inv(0, 1) = -ti[0] / det;
-        jac_inv(1, 0) = -ni[1] / det;
-        jac_inv(1, 1) = ni[0] / det;
-
-        // Fill out the Hessian
-        // (unneeded -- can calculate the inverse components by hand)
-        for (unsigned alpha = 0; alpha < 2; alpha++)
-        {
-          // hess[alpha](0,0) = 0.0;
-          hess[alpha](0, 1) = dnids[alpha];
-          hess[alpha](1, 0) = dnids[alpha];
-          hess[alpha](1, 1) = dtids[alpha];
-        }
-
-        // Fill out inverse of Hessian
-        // H^{-1}abg = J^{-1}ad Hdez J^{-1}eb J^{-1}zg
-        for (unsigned alpha = 0; alpha < 2; alpha++)
-        {
-          for (unsigned beta = 0; beta < 2; beta++)
-          {
-            for (unsigned gamma = 0; gamma < 2; gamma++)
-            {
-              for (unsigned alpha2 = 0; alpha2 < 2; alpha2++)
-              {
-                for (unsigned beta2 = 0; beta2 < 2; beta2++)
-                {
-                  for (unsigned gamma2 = 0; gamma2 < 2; gamma2++)
-                  {
-                    hess_inv[alpha](beta, gamma) -=
-                      jac_inv(alpha, alpha2) * hess[alpha2](beta2, gamma2) *
-                      jac_inv(beta2, beta) * jac_inv(gamma2, gamma);
-                  }
-                }
-              }
-            }
-          }
-        }
-
-        // Fill in the rotation matrix using the new basis
-        fill_in_rotation_matrix_at_node_with_basis(j_node, jac_inv, hess_inv);
+  //       update_rotation_matrices();
+  //     } // end of loop over nodes in node_on_boundary [j]
+  //   } // end of set_nodal_boundary_parametrisation()
 
 
-        // // [zdec] debug
-        // std::ofstream jac_and_hess;
-        // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
-        // jac_and_hess << "Jacobian inverse:" << std::endl
-        // 		   << bi[0][0] << " " << bi[0][1] << std::endl
-        // 		   << bi[1][0] << " " << bi[1][1] << std::endl
-        // 		   << "Hessian inverse [x]:" << std::endl
-        // 		   << Dbi[0](0,0) << " " << Dbi[0](0,1) << std::endl
-        // 		   << Dbi[0](1,0) << " " << Dbi[0](1,1) << std::endl
-        // 		   << "Hessian inverse [y]:" << std::endl
-        // 		   << Dbi[1](0,0) << " " << Dbi[1](0,1) << std::endl
-        // 		   << Dbi[1](1,0) << " " << Dbi[1](1,1) << std::endl <<
-        // std::endl;
+  //   /// Update all rotation matrices (checks if they are needed unless flag is
+  //   /// true)
+  //   void update_rotation_matrices()
+  //   {
+  //     // [zdec] hard coded the three vertex nodes
+  //     unsigned n_vertex = 3;
+  //     // Loop over each vertex
+  //     for (unsigned j_node = 0; j_node < n_vertex; j_node++)
+  //     {
+  //       // If this node does not have a parametrisation (the pointer is still
+  //       // null) skip over it, otherwise we go on to fill out the rotation
+  //       // matrix
+  //       if (!nodal_boundary_parametrisation_pt(j_node))
+  //       {
+  //         continue;
+  //       }
+
+  //       // Initialise the two basis vectors and their jacobians
+  //       Vector<Vector<double>> bi(2, Vector<double>(2, 0.0));
+  //       Vector<DenseMatrix<double>> dbidx(2, DenseMatrix<double>(2, 2, 0.0));
+
+  //       // Our new coordinate system:
+  //       //     (l, s)=(normal component, tangent component)
+  //       // which we define in terms of basis vectors (rescaled)
+  //       //     ni=dxi/dl / |n|           <-- Jacobian col 1
+  //       //     ti=dxi/ds / |t|           <-- Jacobian col 2
+  //       // and their derivatives
+  //       //     dnidxj=d/dxj(dxi/dl / |n|) <-- Hessian `col' 1
+  //       //     dtidxj=d/dxj(dxi/ds / |t|) <-- Hessian `col' 2
+
+  //       // [zdec] we use i and j for brevity
+  //       // but it should be alpha & beta
+  //       // Need to write up how the transformation is done
+
+  //       // Storage for our basis and derivatives
+  //       Vector<double> ni(2, 0.0);
+  //       Vector<double> ti(2, 0.0);
+  //       Vector<double> dnids(2, 0.0);
+  //       Vector<double> dtids(2, 0.0);
+
+  //       // All tensors assumed evaluated on the boundary
+  //       // Jacobian of inverse mapping
+  //       DenseMatrix<double> jac_inv(2, 2, 0.0);
+  //       // Hessian of mapping [zdec] (not needed because...)
+  //       Vector<DenseMatrix<double>> hess(2, DenseMatrix<double>(2, 2, 0.0));
+  //       // Hessian of inverse mapping [zdec] (...this can be found by
+  //       // hand)
+  //       Vector<DenseMatrix<double>> hess_inv(2, DenseMatrix<double>(2, 2, 0.0));
+
+  //       // The basis is defined in terms of the boundary parametrisation
+  //       Vector<double> boundary_coord = {Boundary_coordinate_of_node[j_node]};
+  //       C1CurviLine* boundary_pt =
+  //         Nodal_boundary_parametrisation_pt[j_node];
+  //       Vector<double> x(2, 0.0);
+  //       Vector<double> dxids(2, 0.0);
+  //       Vector<double> d2xids2(2, 0.0);
+
+  //       // Get position (debug) // hierher why debug?
+  //       boundary_pt->position(boundary_coord, x);
+  //       // Get tangent vector
+  //       boundary_pt->dposition(boundary_coord, dxids);
+  //       // Get second derivative
+  //       boundary_pt->d2position(boundary_coord, d2xids2);
+
+  //       double mag_t = sqrt(dxids[0] * dxids[0] + dxids[1] * dxids[1]);
+  //       // ti is the normalised tangent vector
+  //       ti[0] = dxids[0] / mag_t;
+  //       ti[1] = dxids[1] / mag_t;
+  //       // Derivative of (normalised) tangent
+  //       dtids[0] = d2xids2[0] / std::pow(mag_t, 2) -
+  //                  (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[0] /
+  //                    std::pow(mag_t, 4);
+  //       dtids[1] = d2xids2[1] / std::pow(mag_t, 2) -
+  //                  (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[1] /
+  //                    std::pow(mag_t, 4);
+  //       // n = (t x e_z) implies
+  //       ni[0] = ti[1];
+  //       ni[1] = -ti[0];
+  //       // Same for dnids
+  //       dnids[0] = dtids[1];
+  //       dnids[1] = -dtids[0];
+
+  //       // Need inverse of mapping to calculate ds/dxi ----------------
+  //       //   /  dx/dl  dx/ds  \ -1  ___  __1__ /  dy/ds -dx/ds \ .
+  //       //   \  dy/dl  dy/ds  /     ---   det  \ -dy/dl  dx/dl /
+  //       //
+  //       //                          ___  /  dl/dx  dl/dy  \ .
+  //       //                          ---  \  ds/dx  ds/dy  /
+  //       //
+  //       // Fill out inverse of Jacobian
+  //       double det = (ni[0] * ti[1] - ni[1] * ti[0]);
+  //       jac_inv(0, 0) = ti[1] / det;
+  //       jac_inv(0, 1) = -ti[0] / det;
+  //       jac_inv(1, 0) = -ni[1] / det;
+  //       jac_inv(1, 1) = ni[0] / det;
+
+  //       // Fill out the Hessian
+  //       // (unneeded -- can calculate the inverse components by hand)
+  //       for (unsigned alpha = 0; alpha < 2; alpha++)
+  //       {
+  //         // hess[alpha](0,0) = 0.0;
+  //         hess[alpha](0, 1) = dnids[alpha];
+  //         hess[alpha](1, 0) = dnids[alpha];
+  //         hess[alpha](1, 1) = dtids[alpha];
+  //       }
+
+  //       // Fill out inverse of Hessian
+  //       // H^{-1}abg = J^{-1}ad Hdez J^{-1}eb J^{-1}zg
+  //       for (unsigned alpha = 0; alpha < 2; alpha++)
+  //       {
+  //         for (unsigned beta = 0; beta < 2; beta++)
+  //         {
+  //           for (unsigned gamma = 0; gamma < 2; gamma++)
+  //           {
+  //             for (unsigned alpha2 = 0; alpha2 < 2; alpha2++)
+  //             {
+  //               for (unsigned beta2 = 0; beta2 < 2; beta2++)
+  //               {
+  //                 for (unsigned gamma2 = 0; gamma2 < 2; gamma2++)
+  //                 {
+  //                   hess_inv[alpha](beta, gamma) -=
+  //                     jac_inv(alpha, alpha2) * hess[alpha2](beta2, gamma2) *
+  //                     jac_inv(beta2, beta) * jac_inv(gamma2, gamma);
+  //                 }
+  //               }
+  //             }
+  //           }
+  //         }
+  //       }
+
+  //       // Fill in the rotation matrix using the new basis
+  //       fill_in_rotation_matrix_at_node_with_basis(j_node, jac_inv, hess_inv);
 
 
-        // // [zdec] debug
-        // std::ofstream jac_and_hess;
-        // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
-        // jac_and_hess << "Jacobian inverse:" << std::endl
-        //              << jac_inv(0, 0) << " " << jac_inv(0, 1) << std::endl
-        //              << jac_inv(1, 0) << " " << jac_inv(1, 1) << std::endl
-        //              << "Hessian inverse [x]:" << std::endl
-        //              << hess_inv[0](0, 0) << " " << hess_inv[0](0, 1)
-        //              << std::endl
-        //              << hess_inv[0](1, 0) << " " << hess_inv[0](1, 1)
-        //              << std::endl
-        //              << "Hessian inverse [y]:" << std::endl
-        //              << hess_inv[1](0, 0) << " " << hess_inv[1](0, 1)
-        //              << std::endl
-        //              << hess_inv[1](1, 0) << " " << hess_inv[1](1, 1)
-        //              << std::endl
-        //              << std::endl;
-        // jac_and_hess.close();
-
-        // // [zdec] debug
-        // std::ofstream debug_stream;
-        // debug_stream.open("norm_and_tan.dat", std::ios_base::app);
-        // debug_stream << x[0] << " " << x[1] << " " << ni[0] << " " << ni[1]
-        //              << " " << ti[0] << " " << ti[1] << " " << dnids[0] << "
-        //              "
-        //              << dnids[1] << " " << dtids[0] << " " << dtids[1] << " "
-        //              << d2xids2[0] << " " << d2xids2[1] << std::endl;
-        // debug_stream.close();
-
-      } // end loop over vertices
-    } // end of update_rotation_matrices()
+  //       // // [zdec] debug
+  //       // std::ofstream jac_and_hess;
+  //       // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
+  //       // jac_and_hess << "Jacobian inverse:" << std::endl
+  //       // 		   << bi[0][0] << " " << bi[0][1] << std::endl
+  //       // 		   << bi[1][0] << " " << bi[1][1] << std::endl
+  //       // 		   << "Hessian inverse [x]:" << std::endl
+  //       // 		   << Dbi[0](0,0) << " " << Dbi[0](0,1) << std::endl
+  //       // 		   << Dbi[0](1,0) << " " << Dbi[0](1,1) << std::endl
+  //       // 		   << "Hessian inverse [y]:" << std::endl
+  //       // 		   << Dbi[1](0,0) << " " << Dbi[1](0,1) << std::endl
+  //       // 		   << Dbi[1](1,0) << " " << Dbi[1](1,1) << std::endl <<
+  //       // std::endl;
 
 
-    /// Access function to fill out rot_mat using rotation matrix
-    void get_rotation_matrix_at_node(const unsigned& j_node,
-                                     DenseMatrix<double>& rot_mat)
-    {
-      rot_mat = Rotation_matrix_at_node[j_node];
-    }
+  //       // // [zdec] debug
+  //       // std::ofstream jac_and_hess;
+  //       // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
+  //       // jac_and_hess << "Jacobian inverse:" << std::endl
+  //       //              << jac_inv(0, 0) << " " << jac_inv(0, 1) << std::endl
+  //       //              << jac_inv(1, 0) << " " << jac_inv(1, 1) << std::endl
+  //       //              << "Hessian inverse [x]:" << std::endl
+  //       //              << hess_inv[0](0, 0) << " " << hess_inv[0](0, 1)
+  //       //              << std::endl
+  //       //              << hess_inv[0](1, 0) << " " << hess_inv[0](1, 1)
+  //       //              << std::endl
+  //       //              << "Hessian inverse [y]:" << std::endl
+  //       //              << hess_inv[1](0, 0) << " " << hess_inv[1](0, 1)
+  //       //              << std::endl
+  //       //              << hess_inv[1](1, 0) << " " << hess_inv[1](1, 1)
+  //       //              << std::endl
+  //       //              << std::endl;
+  //       // jac_and_hess.close();
 
-  private:
-    /// Helper function to fill in the rotation matrix for a given basis
-    void fill_in_rotation_matrix_at_node_with_basis(
-      const unsigned& j_node,
-      const DenseMatrix<double>& jac_inv,
-      const Vector<DenseMatrix<double>>& hess_inv)
-    {
-      // Rotation matrix, b constructed using submatrices b1, b12, b22
-      DenseMatrix<double> b1(2, 2, 0.0), b22(3, 3, 0.0), b12(2, 3, 0.0);
+  //       // // [zdec] debug
+  //       // std::ofstream debug_stream;
+  //       // debug_stream.open("norm_and_tan.dat", std::ios_base::app);
+  //       // debug_stream << x[0] << " " << x[1] << " " << ni[0] << " " << ni[1]
+  //       //              << " " << ti[0] << " " << ti[1] << " " << dnids[0] << "
+  //       //              "
+  //       //              << dnids[1] << " " << dtids[0] << " " << dtids[1] << " "
+  //       //              << d2xids2[0] << " " << d2xids2[1] << std::endl;
+  //       // debug_stream.close();
 
-      // Fill in the submatrices
-      // Loop over the rotated first derivatives
-      for (unsigned mu = 0; mu < 2; mu++)
-      {
-        // Loop over the unrotated first derivatives
-        for (unsigned alpha = 0; alpha < 2; alpha++)
-        {
-          // Fill in b1 - the Jacobian
-          // Fill in the affine rotation of the first derivatives
-          b1(mu, alpha) = jac_inv(mu, alpha);
+  //     } // end loop over vertices
+  //   } // end of update_rotation_matrices()
 
-          // Loop over unrotated second derivatives
-          for (unsigned beta = 0; beta < 2; ++beta)
-          {
-            // Avoid double counting the cross derivative
-            if (alpha <= beta)
-            {
-              // Define column index
-              const unsigned col = alpha + beta;
 
-              // Fill in the non-affine part of the rotation of the first
-              // derivatives
-              b12(mu, col) += hess_inv[mu](alpha, beta);
-              // [zdec] debug mixed derivative -- add extra
-              if (alpha < beta)
-              {
-                // b12(mu, col) -= hess_inv[mu](alpha, beta);
-              }
-              // Loop over the rotated second derivatives
-              for (unsigned nu = 0; nu < 2; nu++)
-              {
-                // // Avoid double counting the cross derivative
-                // if (mu <= nu)
-                {
-                  // Fill in b22 - the Affine part of the Jacobian derivative
-                  // Redefine row index for the next submatrix
-                  unsigned row_b22 = mu + nu;
-                  // Fill in the affine part of the rotation of the second
-                  // derivatives [zdec] if( beta>= alpha) ?
-                  b22(row_b22, col) += jac_inv(mu, alpha) * jac_inv(nu, beta);
-                }
-              }
-            }
-          }
-        }
-      }
+  //   /// Access function to fill out rot_mat using rotation matrix
+  //   void get_rotation_matrix_at_node(const unsigned& j_node,
+  //                                    DenseMatrix<double>& rot_mat)
+  //   {
+  //     rot_mat = Rotation_matrix_at_node[j_node];
+  //   }
 
-      // Fill in the submatrices to the full (6x6) matrix
-      Rotation_matrix_at_node[j_node](0, 0) = 1.0;
-      // Fill in b1 --- the affine contribution to rotation of the
-      // first derivatives
-      for (unsigned i = 0; i < 2; ++i)
-      {
-        for (unsigned j = 0; j < 2; ++j)
-        {
-          Rotation_matrix_at_node[j_node](1 + i, 1 + j) = b1(i, j);
-        }
-      }
-      // Fill in b21 --- the non-affine (second derivative dependent)
-      // rotation of the first derivatives
-      for (unsigned i = 0; i < 2; ++i)
-      {
-        for (unsigned j = 0; j < 3; ++j)
-        {
-          Rotation_matrix_at_node[j_node](1 + i, 3 + j) = b12(i, j);
-        }
-      }
-      // Fill in b22 --- the rotation of the second derivatives
-      for (unsigned i = 0; i < 3; ++i)
-      {
-        for (unsigned j = 0; j < 3; ++j)
-        {
-          Rotation_matrix_at_node[j_node](3 + i, 3 + j) = b22(i, j);
-        }
-      }
-    } // end fill_in_rotation_matrix_at_node_with_basis
+  // private:
+  //   /// Helper function to fill in the rotation matrix for a given basis
+  //   void fill_in_rotation_matrix_at_node_with_basis(
+  //     const unsigned& j_node,
+  //     const DenseMatrix<double>& jac_inv,
+  //     const Vector<DenseMatrix<double>>& hess_inv)
+  //   {
+  //     // Rotation matrix, b constructed using submatrices b1, b12, b22
+  //     DenseMatrix<double> b1(2, 2, 0.0), b22(3, 3, 0.0), b12(2, 3, 0.0);
 
-    /// Pointer to the `parent' finite element which this is a helper force
-    FiniteElement* Parent_element_pt;
+  //     // Fill in the submatrices
+  //     // Loop over the rotated first derivatives
+  //     for (unsigned mu = 0; mu < 2; mu++)
+  //     {
+  //       // Loop over the unrotated first derivatives
+  //       for (unsigned alpha = 0; alpha < 2; alpha++)
+  //       {
+  //         // Fill in b1 - the Jacobian
+  //         // Fill in the affine rotation of the first derivatives
+  //         b1(mu, alpha) = jac_inv(mu, alpha);
 
-    /// The number of nodes (that we store rotation data for) in the fvk element
-    /// that uses this helper
-    unsigned Nnode;
+  //         // Loop over unrotated second derivatives
+  //         for (unsigned beta = 0; beta < 2; ++beta)
+  //         {
+  //           // Avoid double counting the cross derivative
+  //           if (alpha <= beta)
+  //           {
+  //             // Define column index
+  //             const unsigned col = alpha + beta;
 
-    /// Vector containing boundary parametrised location for each node
-    Vector<double> Boundary_coordinate_of_node;
+  //             // Fill in the non-affine part of the rotation of the first
+  //             // derivatives
+  //             b12(mu, col) += hess_inv[mu](alpha, beta);
+  //             // [zdec] debug mixed derivative -- add extra
+  //             if (alpha < beta)
+  //             {
+  //               // b12(mu, col) -= hess_inv[mu](alpha, beta);
+  //             }
+  //             // Loop over the rotated second derivatives
+  //             for (unsigned nu = 0; nu < 2; nu++)
+  //             {
+  //               // // Avoid double counting the cross derivative
+  //               // if (mu <= nu)
+  //               {
+  //                 // Fill in b22 - the Affine part of the Jacobian derivative
+  //                 // Redefine row index for the next submatrix
+  //                 unsigned row_b22 = mu + nu;
+  //                 // Fill in the affine part of the rotation of the second
+  //                 // derivatives [zdec] if( beta>= alpha) ?
+  //                 b22(row_b22, col) += jac_inv(mu, alpha) * jac_inv(nu, beta);
+  //               }
+  //             }
+  //           }
+  //         }
+  //       }
+  //     }
 
-    /// Vector containing boundary parametrisation at each node
-    Vector<CurvilineGeomObject*> Nodal_boundary_parametrisation_pt;
+  //     // Fill in the submatrices to the full (6x6) matrix
+  //     Rotation_matrix_at_node[j_node](0, 0) = 1.0;
+  //     // Fill in b1 --- the affine contribution to rotation of the
+  //     // first derivatives
+  //     for (unsigned i = 0; i < 2; ++i)
+  //     {
+  //       for (unsigned j = 0; j < 2; ++j)
+  //       {
+  //         Rotation_matrix_at_node[j_node](1 + i, 1 + j) = b1(i, j);
+  //       }
+  //     }
+  //     // Fill in b21 --- the non-affine (second derivative dependent)
+  //     // rotation of the first derivatives
+  //     for (unsigned i = 0; i < 2; ++i)
+  //     {
+  //       for (unsigned j = 0; j < 3; ++j)
+  //       {
+  //         Rotation_matrix_at_node[j_node](1 + i, 3 + j) = b12(i, j);
+  //       }
+  //     }
+  //     // Fill in b22 --- the rotation of the second derivatives
+  //     for (unsigned i = 0; i < 3; ++i)
+  //     {
+  //       for (unsigned j = 0; j < 3; ++j)
+  //       {
+  //         Rotation_matrix_at_node[j_node](3 + i, 3 + j) = b22(i, j);
+  //       }
+  //     }
+  //   } // end fill_in_rotation_matrix_at_node_with_basis
 
-    /// Vector containing <rotation matrix at each node>
-    Vector<DenseMatrix<double>> Rotation_matrix_at_node;
-  };
-  //---end of rotation helper class-------------------------------------------
+  //   /// Pointer to the `parent' finite element which this is a helper force
+  //   FiniteElement* Parent_element_pt;
+
+  //   /// The number of nodes (that we store rotation data for) in the fvk element
+  //   /// that uses this helper
+  //   unsigned Nnode;
+
+  //   /// Vector containing boundary parametrised location for each node
+  //   Vector<double> Boundary_coordinate_of_node;
+
+  //   /// Vector containing boundary parametrisation at each node
+  //   Vector<C1CurviLine*> Nodal_boundary_parametrisation_pt;
+
+  //   /// Vector containing <rotation matrix at each node>
+  //   Vector<DenseMatrix<double>> Rotation_matrix_at_node;
+  // };
+  // //---end of rotation helper class-------------------------------------------
 
 
   //============================================================================
@@ -619,22 +621,23 @@ namespace oomph
       return CurvableBellElement<NNODE_1D>::element_is_curved();
     }
 
-    /// Upgrade the Bell element to a curved Bernadou element. Expects, in
-    /// order, the unsigned enumeration of the edge that is on the boundary
-    /// (curved_edge) as well as the coordinates of the start and end of that
-    /// edge on the boundary (s_ubar,s_obar), the parametric description of the
-    /// curved edge (parametric_edge) and lastly the polynomial order of the
-    /// boundary interpolation (boundary_order) which can be either 3 or 5.
-    virtual void upgrade_element_to_curved(
-      const MyC1CurvedElements::Edge& curved_edge,
-      const double& s_ubar,
-      const double& s_obar,
-      CurvilineGeomObject* parametric_edge,
-      const unsigned& boundary_order)
-    {
-      CurvableBellElement<NNODE_1D>::upgrade_element_to_curved(
-        curved_edge, s_ubar, s_obar, parametric_edge, boundary_order);
-    }
+   // hierher why doesn't this just call the underlying function by inheritance? no wrapper needed? 
+    // /// Upgrade the Bell element to a curved Bernadou element. Expects, in
+    // /// order, the unsigned enumeration of the edge that is on the boundary
+    // /// (curved_edge) as well as the coordinates of the start and end of that
+    // /// edge on the boundary (s_ubar,s_obar), the parametric description of the
+    // /// curved edge (parametric_edge) and lastly the polynomial order of the
+    // /// boundary interpolation (boundary_order) which can be either 3 or 5.
+    // virtual void upgrade_element_to_curved(
+    //   const MyC1CurvedElements::Edge& curved_edge,
+    //   const double& s_ubar,
+    //   const double& s_obar,
+    //   C1CurviLine* parametric_edge,
+    //   const unsigned& boundary_order)
+    // {
+    //   CurvableBellElement<NNODE_1D>::upgrade_element_to_curved(
+    //     curved_edge, s_ubar, s_obar, parametric_edge, boundary_order);
+    // }
 
 
     //----------------------------------------------------------------------
@@ -1817,8 +1820,8 @@ of freedom at internal points. They are {w ; w,x ; w,y ; w,xx ; w,xy ; w,yy}",
     DuplicateNodeConstraintElement(
       Node* const& left_node_pt,
       Node* const& right_node_pt,
-      CurvilineGeomObject* const& left_boundary_pt,
-      CurvilineGeomObject* const& right_boundary_pt,
+      C1CurviLine* const& left_boundary_pt,
+      C1CurviLine* const& right_boundary_pt,
       Vector<double> const& left_coord,
       Vector<double> const& right_coord)
       : Left_node_pt(left_node_pt),
@@ -2229,7 +2232,7 @@ of freedom at internal points. They are {w ; w,x ; w,y ; w,xx ; w,xy ; w,yy}",
       Vector<double> left_d2xids2(2, 0.0);
       Vector<double> right_dxids(2, 0.0);
       Vector<double> right_d2xids2(2, 0.0);
-      Left_boundary_pt->position(Left_node_coord, left_x); // [zdec] debug
+      Left_boundary_pt->position(Left_node_coord, left_x); // [zdec] debug // hierher why debug?
       Right_boundary_pt->position(Right_node_coord, right_x); // [zdec] debug
       Left_boundary_pt->dposition(Left_node_coord, left_dxids);
       Left_boundary_pt->d2position(Left_node_coord, left_d2xids2);
@@ -2849,10 +2852,10 @@ of freedom at internal points. They are {w ; w,x ; w,y ; w,xx ; w,xy ; w,yy}",
     Node* Right_node_pt;
 
     /// Pointer to the left node's boundary parametrisation
-    CurvilineGeomObject* Left_boundary_pt;
+    C1CurviLine* Left_boundary_pt;
 
     /// Pointer to the right node's boundary parametrisation
-    CurvilineGeomObject* Right_boundary_pt;
+    C1CurviLine* Right_boundary_pt;
 
     /// Coordinate of the left node on the left boundary
     Vector<double> Left_node_coord;
