+++ src/generic/my_geom_object.h
@@ -27,690 +27,519 @@
 // LIC// The authors may be contacted at oomph-lib@maths.man.ac.uk.
 // LIC//
 // LIC//====================================================================
-#ifndef OOMPH_MY_GEOM_OBJECTS_HEADER
+#ifndef OOMPH_MY_GEOM_OBJECTS_HEADER // hierher C1_curviline.h
 #define OOMPH_MY_GEOM_OBJECTS_HEADER
-#include "geom_objects.h"
+
+#include "unstructured_two_d_mesh_geometry_base.h"
+
+
 namespace oomph
 {
-  ////////////////////////////////////////////////////////////////////////
-  ////////////////////////////////////////////////////////////////////////
-  // My Geometric object
-  ////////////////////////////////////////////////////////////////////////
-  ////////////////////////////////////////////////////////////////////////
-  /// Specialisation of GeomObject that has a scalar parametric zeta and
-  /// a 2D vector r coordinate r = (r_1(zeta),r_2(zeta)).
-  /// This is suitable for use in meshing/construction of planar objects.
-  class CurvilineGeomObject : public GeomObject
+
+ 
+  //======================================================================
+  /// hierher
+  //======================================================================
+  class C1CurviLine 
 
   {
   public:
-    /// Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineGeomObject() : GeomObject(1, 2) {}
-
-    /// Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineGeomObject(TimeStepper* time_stepper_pt)
-      : GeomObject(1, 2, time_stepper_pt)
+   
+   /// Constructor: pointer to underlying TriangleMeshCurviline object
+   C1CurviLine(TriangleMeshCurviLine* triangle_mesh_curviline_pt) :
+    Triangle_mesh_curviline_pt(triangle_mesh_curviline_pt)
     {
     }
-
+   
     /// Broken copy constructor
-    CurvilineGeomObject(const CurvilineGeomObject& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineGeomObject");
-    }
+   C1CurviLine(const C1CurviLine& dummy) = delete;
 
     /// Broken assignment operator
-    void operator=(const CurvilineGeomObject&)
+   void operator=(const C1CurviLine&) = delete;
+   
+    /// (Empty) destructor
+    virtual ~C1CurviLine() {}
+
+
+   // hierher in all of these, why is zeta still a vector?
+   
+   /// Position r as fct of zeta; forward to underlying geom object
+   void position(const Vector<double>& zeta,
+                 Vector<double>& r) const
     {
-      BrokenCopy::broken_assign("CurvilineGeomObject");
+     Triangle_mesh_curviline_pt->geom_object_pt()->position(zeta,r); 
     }
 
-    /// (Empty) destructor
-    virtual ~CurvilineGeomObject() {}
-
+   
     /// Derivative of position Vector w.r.t. to zeta:
     virtual void dposition(const Vector<double>& zeta,
                            Vector<double>& drdzeta) const
     {
-      throw OomphLibError(
-        "You must specify dposition() for your own object! \n",
-        OOMPH_CURRENT_FUNCTION,
-        OOMPH_EXCEPTION_LOCATION);
+     DenseMatrix<double> drdzeta_general(2,2);
+     Triangle_mesh_curviline_pt->geom_object_pt()->dposition(zeta,drdzeta_general);
+     drdzeta[0]=drdzeta_general(0,0);
+     drdzeta[1]=drdzeta_general(0,1);
     }
-
-
-    /// 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
+   
+   /// 2nd derivative of position Vector w.r.t. to coordinates:
+   /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
+   /// ddrdzeta(alpha,beta,i).
+   /// Evaluated at current time.
+   virtual void d2position(const Vector<double>& zeta,
+                           Vector<double>& d2rdzeta) const
     {
-      throw OomphLibError(
-        "You must specify d2position() for your own object! \n",
-        OOMPH_CURRENT_FUNCTION,
-        OOMPH_EXCEPTION_LOCATION);
+     RankThreeTensor<double> d2rdzeta_general(1,1,2);
+     Triangle_mesh_curviline_pt->geom_object_pt()->d2position(zeta,d2rdzeta_general);
+     d2rdzeta[0]=d2rdzeta_general(0,0,0);
+     d2rdzeta[1]=d2rdzeta_general(0,0,1);
+    }
+   
+   
+   /// The underlying GeomObject encodes r(zeta); here we invert this mapping
+   /// with a tolerance to get zeta associated with the given point r_target.
+   /// Default implemenation based on bisection and Newton's method.
+   /// Overload with your own!
+   virtual double get_zeta(const Vector<double>& r_target,
+                           const double& tol=1.0e-8) const
+    {
+     Vector<double> r(2);
+     Vector<double> zeta_vect(1);
+     Vector<double> drdzeta(2);
+     
+     // Start Newton method in the middle
+     double zeta_min=Triangle_mesh_curviline_pt->zeta_start();
+     double zeta_max=Triangle_mesh_curviline_pt->zeta_end();
+     double zeta=0.5*(zeta_max+zeta_min);
+
+     // Do it
+     unsigned max_iter=100;
+     for (unsigned iter=0;iter<max_iter;iter++)
+      {
+       // Residual:
+       zeta_vect[0]=zeta;
+       position(zeta_vect,r);
+       double res=sqrt( (r[0]-r_target[0])*(r[0]-r_target[0]) +
+                        (r[1]-r_target[1])*(r[1]-r_target[1]) );
+       if (res<tol)
+        {
+         if ( (zeta>zeta_min-tol) &&
+              (zeta<zeta_max+tol) )
+          {
+           oomph_info << "Return with zeta = " << zeta << std::endl;
+           return zeta;
+          }
+         else
+          {
+           oomph_info << "Converged to zeta outside range: zeta = "
+                      << zeta << std::endl
+                      << "zeta_min/max = " << zeta_min << " " << zeta_max
+                      << std::endl;
+           abort(); // hierher
+          }
+        }
+       
+       oomph_info << "iter, zeta, res "
+                  << iter << " "
+                  << zeta << " "
+                  << res << " "
+                  << std::endl;
+       
+       // Get "Jacobian"
+       dposition(zeta_vect,drdzeta);
+       double dresdzeta=
+        ( (r[0]-r_target[0])*drdzeta[0] +
+          (r[1]-r_target[1])*drdzeta[1] )/res;
+
+       // Newton correctino
+       zeta-=res/dresdzeta;
+      }
+
+     // If we get here died:
+     oomph_info << "newton method failed to converge after max_iter = "
+                << max_iter << " iterations" << std::endl;
+     abort(); // hierher
+
+     // dummy return
+     return zeta;
     }
 
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    virtual double get_zeta(const Vector<double>& x) const
-    {
-      throw OomphLibError("You must specify get_zeta() for your own object! \n",
-                          OOMPH_CURRENT_FUNCTION,
-                          OOMPH_EXCEPTION_LOCATION);
-    }
+
+   // hierher may retain but all functions should have been wrapped now.
+   // /// Pointer to underlying triangle object
+   // TriangleMeshCurviLine* triangle_mesh_curviline_pt() const
+   //  {
+   //   return Triangle_mesh_curviline_pt;
+   //  }
+   
+
+  private:
+
+   /// Pointer to underlying triangle object
+   TriangleMeshCurviLine* Triangle_mesh_curviline_pt;
+   
   };
 
 
-  /// Specialisation of CurvilineGeomObject for a straight line
-  class CurvilineLine : public CurvilineGeomObject
+
+
+
+////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////
+
+
+
+  //===start of rotation helper class=========================================
+  /// Helper class to contain all the rotation information in the element.
+  class RotatedBoundaryHelper
   {
   public:
-    /// Constructor: Unit line between origin and (1 , 0)
-    CurvilineLine() : CurvilineGeomObject(), X0{0.0, 0.0}, X1{1.0, 0.0} {}
+    /// Constructor: just initialise the member data to their defaults (zeros)
+    RotatedBoundaryHelper(FiniteElement* const& parent_element_pt)
+      : Parent_element_pt(parent_element_pt),
+        Nnode(Parent_element_pt->nvertex_node()),
+        Boundary_coordinate_of_node(3, 0.0),
+        Nodal_boundary_parametrisation_pt(3, 0),
+        Rotation_matrix_at_node(3, DenseMatrix<double>(6, 6, 0.0))
+    {
+    }
 
-    /// Constructor: Unit radius from origin at angle theta. The reverse flag
-    /// flips the parametrisation direction. Useful for when only the slope
-    /// matters.
-    CurvilineLine(const double& theta, const bool& reverse = false)
-      : CurvilineGeomObject(), X0{0.0, 0.0}, X1{cos(theta), sin(theta)}
-    { // Flip the parametrisation if requested
-      if (reverse)
+    /// Destructor
+    ~RotatedBoundaryHelper() {}
+
+    C1CurviLine* nodal_boundary_parametrisation_pt(
+      const unsigned& j_node)
+    {
+      return Nodal_boundary_parametrisation_pt[j_node];
+    }
+
+   // hierher isn't this the same as in fkv?
+
+    /// Add a new boundary parametrisation to nodes all the nodes in the
+    /// vector node_on_boundary
+    void set_nodal_boundary_parametrisation(
+      const Vector<unsigned>& node_on_boundary,
+      const Vector<double>& boundary_coord_of_node,
+      C1CurviLine* const& boundary_parametrisation_pt)
+    {
+      // Loop over all the nodes in node_on_boundary and add the boundary
+      // pointer to their vector of boundaries
+      unsigned n_node = node_on_boundary.size();
+      for (unsigned j = 0; j < n_node; j++)
       {
-        Vector<double> temp = X1;
-        X1[0] = X0[0];
-        X1[1] = X0[1];
-        X0[0] = temp[0];
-        X0[1] = temp[1];
-      }
-    }
+        // The j-th node on the boundary
+        unsigned j_node = node_on_boundary[j];
 
-    /// Constructor: Straight line between x0 and x1
-    CurvilineLine(const Vector<double>& x0, const Vector<double>& x1)
-      : CurvilineGeomObject(), X0(x0), X1(x1)
+        // Set the boundary parametrisation data pointer for this node
+        Nodal_boundary_parametrisation_pt[j_node] = boundary_parametrisation_pt;
+
+        // Set the coordinate of node j on this boundary
+        Boundary_coordinate_of_node[j_node] = boundary_coord_of_node[j];
+
+        update_rotation_matrices();
+      } // end of loop over nodes in node_on_boundary [j]
+    } // end of set_nodal_boundary_parametrisation()
+
+
+    /// Update all rotation matrices (checks if they are needed unless flag is
+    /// true)
+    void update_rotation_matrices()
     {
-      // Do nothing
-    }
+      // [zdec] hard coded the three vertex nodes
+      unsigned n_vertex = 3;
+      // Loop over each vertex
+      for (unsigned j_node = 0; j_node < n_vertex; j_node++)
+      {
+        // If this node does not have a parametrisation (the pointer is still
+        // null) skip over it, otherwise we go on to fill out the rotation
+        // matrix
+        if (!nodal_boundary_parametrisation_pt(j_node))
+        {
+          continue;
+        }
 
-    /// Broken copy constructor
-    CurvilineLine(const CurvilineLine& dummy)
+        // Initialise the two basis vectors and their jacobians
+        Vector<Vector<double>> bi(2, Vector<double>(2, 0.0));
+        Vector<DenseMatrix<double>> dbidx(2, DenseMatrix<double>(2, 2, 0.0));
+
+        // Our new coordinate system:
+        //     (l, s)=(normal component, tangent component)
+        // which we define in terms of basis vectors (rescaled)
+        //     ni=dxi/dl / |n|           <-- Jacobian col 1
+        //     ti=dxi/ds / |t|           <-- Jacobian col 2
+        // and their derivatives
+        //     dnidxj=d/dxj(dxi/dl / |n|) <-- Hessian `col' 1
+        //     dtidxj=d/dxj(dxi/ds / |t|) <-- Hessian `col' 2
+
+        // [zdec] we use i and j for brevity
+        // but it should be alpha & beta
+        // Need to write up how the transformation is done
+
+        // Storage for our basis and derivatives
+        Vector<double> ni(2, 0.0);
+        Vector<double> ti(2, 0.0);
+        Vector<double> dnids(2, 0.0);
+        Vector<double> dtids(2, 0.0);
+
+        // All tensors assumed evaluated on the boundary
+        // Jacobian of inverse mapping
+        DenseMatrix<double> jac_inv(2, 2, 0.0);
+        // Hessian of mapping [zdec] (not needed because...)
+        Vector<DenseMatrix<double>> hess(2, DenseMatrix<double>(2, 2, 0.0));
+        // Hessian of inverse mapping [zdec] (...this can be found by
+        // hand)
+        Vector<DenseMatrix<double>> hess_inv(2, DenseMatrix<double>(2, 2, 0.0));
+
+        // The basis is defined in terms of the boundary parametrisation
+        Vector<double> boundary_coord = {Boundary_coordinate_of_node[j_node]};
+        C1CurviLine* boundary_pt =
+          Nodal_boundary_parametrisation_pt[j_node];
+        Vector<double> x(2, 0.0);
+        Vector<double> dxids(2, 0.0);
+        Vector<double> d2xids2(2, 0.0);
+
+        // Get position (debug) // hierher why debug?
+        boundary_pt->position(boundary_coord, x);
+        // Get tangent vector
+        boundary_pt->dposition(boundary_coord, dxids);
+        // Get second derivative
+        boundary_pt->d2position(boundary_coord, d2xids2);
+
+        double mag_t = sqrt(dxids[0] * dxids[0] + dxids[1] * dxids[1]);
+        // ti is the normalised tangent vector
+        ti[0] = dxids[0] / mag_t;
+        ti[1] = dxids[1] / mag_t;
+        // Derivative of (normalised) tangent
+        dtids[0] = d2xids2[0] / std::pow(mag_t, 2) -
+                   (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[0] /
+                     std::pow(mag_t, 4);
+        dtids[1] = d2xids2[1] / std::pow(mag_t, 2) -
+                   (dxids[0] * d2xids2[0] + dxids[1] * d2xids2[1]) * dxids[1] /
+                     std::pow(mag_t, 4);
+        // n = (t x e_z) implies
+        ni[0] = ti[1];
+        ni[1] = -ti[0];
+        // Same for dnids
+        dnids[0] = dtids[1];
+        dnids[1] = -dtids[0];
+
+        // Need inverse of mapping to calculate ds/dxi ----------------
+        //   /  dx/dl  dx/ds  \ -1  ___  __1__ /  dy/ds -dx/ds \ .
+        //   \  dy/dl  dy/ds  /     ---   det  \ -dy/dl  dx/dl /
+        //
+        //                          ___  /  dl/dx  dl/dy  \ .
+        //                          ---  \  ds/dx  ds/dy  /
+        //
+        // Fill out inverse of Jacobian
+        double det = (ni[0] * ti[1] - ni[1] * ti[0]);
+        jac_inv(0, 0) = ti[1] / det;
+        jac_inv(0, 1) = -ti[0] / det;
+        jac_inv(1, 0) = -ni[1] / det;
+        jac_inv(1, 1) = ni[0] / det;
+
+        // Fill out the Hessian
+        // (unneeded -- can calculate the inverse components by hand)
+        for (unsigned alpha = 0; alpha < 2; alpha++)
+        {
+          // hess[alpha](0,0) = 0.0;
+          hess[alpha](0, 1) = dnids[alpha];
+          hess[alpha](1, 0) = dnids[alpha];
+          hess[alpha](1, 1) = dtids[alpha];
+        }
+
+        // Fill out inverse of Hessian
+        // H^{-1}abg = J^{-1}ad Hdez J^{-1}eb J^{-1}zg
+        for (unsigned alpha = 0; alpha < 2; alpha++)
+        {
+          for (unsigned beta = 0; beta < 2; beta++)
+          {
+            for (unsigned gamma = 0; gamma < 2; gamma++)
+            {
+              for (unsigned alpha2 = 0; alpha2 < 2; alpha2++)
+              {
+                for (unsigned beta2 = 0; beta2 < 2; beta2++)
+                {
+                  for (unsigned gamma2 = 0; gamma2 < 2; gamma2++)
+                  {
+                    hess_inv[alpha](beta, gamma) -=
+                      jac_inv(alpha, alpha2) * hess[alpha2](beta2, gamma2) *
+                      jac_inv(beta2, beta) * jac_inv(gamma2, gamma);
+                  }
+                }
+              }
+            }
+          }
+        }
+
+        // Fill in the rotation matrix using the new basis
+        fill_in_rotation_matrix_at_node_with_basis(j_node, jac_inv, hess_inv);
+
+
+        // // [zdec] debug
+        // std::ofstream jac_and_hess;
+        // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
+        // jac_and_hess << "Jacobian inverse:" << std::endl
+        // 		   << bi[0][0] << " " << bi[0][1] << std::endl
+        // 		   << bi[1][0] << " " << bi[1][1] << std::endl
+        // 		   << "Hessian inverse [x]:" << std::endl
+        // 		   << Dbi[0](0,0) << " " << Dbi[0](0,1) << std::endl
+        // 		   << Dbi[0](1,0) << " " << Dbi[0](1,1) << std::endl
+        // 		   << "Hessian inverse [y]:" << std::endl
+        // 		   << Dbi[1](0,0) << " " << Dbi[1](0,1) << std::endl
+        // 		   << Dbi[1](1,0) << " " << Dbi[1](1,1) << std::endl <<
+        // std::endl;
+
+
+        // // [zdec] debug
+        // std::ofstream jac_and_hess;
+        // jac_and_hess.open("jac_and_hess_new.csv", std::ios_base::app);
+        // jac_and_hess << "Jacobian inverse:" << std::endl
+        //              << jac_inv(0, 0) << " " << jac_inv(0, 1) << std::endl
+        //              << jac_inv(1, 0) << " " << jac_inv(1, 1) << std::endl
+        //              << "Hessian inverse [x]:" << std::endl
+        //              << hess_inv[0](0, 0) << " " << hess_inv[0](0, 1)
+        //              << std::endl
+        //              << hess_inv[0](1, 0) << " " << hess_inv[0](1, 1)
+        //              << std::endl
+        //              << "Hessian inverse [y]:" << std::endl
+        //              << hess_inv[1](0, 0) << " " << hess_inv[1](0, 1)
+        //              << std::endl
+        //              << hess_inv[1](1, 0) << " " << hess_inv[1](1, 1)
+        //              << std::endl
+        //              << std::endl;
+        // jac_and_hess.close();
+
+        // // [zdec] debug
+        // std::ofstream debug_stream;
+        // debug_stream.open("norm_and_tan.dat", std::ios_base::app);
+        // debug_stream << x[0] << " " << x[1] << " " << ni[0] << " " << ni[1]
+        //              << " " << ti[0] << " " << ti[1] << " " << dnids[0] << "
+        //              "
+        //              << dnids[1] << " " << dtids[0] << " " << dtids[1] << " "
+        //              << d2xids2[0] << " " << d2xids2[1] << std::endl;
+        // debug_stream.close();
+
+      } // end loop over vertices
+    } // end of update_rotation_matrices()
+
+
+    /// Access function to fill out rot_mat using rotation matrix
+    void get_rotation_matrix_at_node(const unsigned& j_node,
+                                     DenseMatrix<double>& rot_mat)
     {
-      BrokenCopy::broken_copy("CurvilineLine");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineLine&)
-    {
-      BrokenCopy::broken_assign("CurvilineLine");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineLine() {}
-
-    /// Access function to set the start of the line
-    void set_ends(const Vector<double>& x0, const Vector<double>& x1)
-    {
-      X0 = x0;
-      X1 = x1;
-    }
-
-    /// Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = (1.0 - zeta[0]) * X0[0] + zeta[0] * X1[0];
-      r[1] = (1.0 - zeta[0]) * X0[1] + zeta[0] * X1[1];
-    }
-
-    /// Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = X1[0] - X0[0];
-      drdzeta[1] = X1[1] - X0[1];
-    }
-
-
-    /// 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = 0.0;
-      drdzeta[1] = 0.0;
-    }
-
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The parametric parameter
-      return (X1[0] - X0[0] != 0 ? (x[0] - X0[0]) / (X1[0] - X0[0]) :
-                                   (x[1] - X0[1]) / (X1[1] - X0[1]));
+      rot_mat = Rotation_matrix_at_node[j_node];
     }
 
   private:
-    /// Start coordinates
-    Vector<double> X0;
+    /// Helper function to fill in the rotation matrix for a given basis
+    void fill_in_rotation_matrix_at_node_with_basis(
+      const unsigned& j_node,
+      const DenseMatrix<double>& jac_inv,
+      const Vector<DenseMatrix<double>>& hess_inv)
+    {
+      // Rotation matrix, b constructed using submatrices b1, b12, b22
+      DenseMatrix<double> b1(2, 2, 0.0), b22(3, 3, 0.0), b12(2, 3, 0.0);
 
-    /// End coordinates
-    Vector<double> X1;
+      // Fill in the submatrices
+      // Loop over the rotated first derivatives
+      for (unsigned mu = 0; mu < 2; mu++)
+      {
+        // Loop over the unrotated first derivatives
+        for (unsigned alpha = 0; alpha < 2; alpha++)
+        {
+          // Fill in b1 - the Jacobian
+          // Fill in the affine rotation of the first derivatives
+          b1(mu, alpha) = jac_inv(mu, alpha);
+
+          // Loop over unrotated second derivatives
+          for (unsigned beta = 0; beta < 2; ++beta)
+          {
+            // Avoid double counting the cross derivative
+            if (alpha <= beta)
+            {
+              // Define column index
+              const unsigned col = alpha + beta;
+
+              // Fill in the non-affine part of the rotation of the first
+              // derivatives
+              b12(mu, col) += hess_inv[mu](alpha, beta);
+              // [zdec] debug mixed derivative -- add extra
+              if (alpha < beta)
+              {
+                // b12(mu, col) -= hess_inv[mu](alpha, beta);
+              }
+              // Loop over the rotated second derivatives
+              for (unsigned nu = 0; nu < 2; nu++)
+              {
+                // // Avoid double counting the cross derivative
+                // if (mu <= nu)
+                {
+                  // Fill in b22 - the Affine part of the Jacobian derivative
+                  // Redefine row index for the next submatrix
+                  unsigned row_b22 = mu + nu;
+                  // Fill in the affine part of the rotation of the second
+                  // derivatives [zdec] if( beta>= alpha) ?
+                  b22(row_b22, col) += jac_inv(mu, alpha) * jac_inv(nu, beta);
+                }
+              }
+            }
+          }
+        }
+      }
+
+      // Fill in the submatrices to the full (6x6) matrix
+      Rotation_matrix_at_node[j_node](0, 0) = 1.0;
+      // Fill in b1 --- the affine contribution to rotation of the
+      // first derivatives
+      for (unsigned i = 0; i < 2; ++i)
+      {
+        for (unsigned j = 0; j < 2; ++j)
+        {
+          Rotation_matrix_at_node[j_node](1 + i, 1 + j) = b1(i, j);
+        }
+      }
+      // Fill in b21 --- the non-affine (second derivative dependent)
+      // rotation of the first derivatives
+      for (unsigned i = 0; i < 2; ++i)
+      {
+        for (unsigned j = 0; j < 3; ++j)
+        {
+          Rotation_matrix_at_node[j_node](1 + i, 3 + j) = b12(i, j);
+        }
+      }
+      // Fill in b22 --- the rotation of the second derivatives
+      for (unsigned i = 0; i < 3; ++i)
+      {
+        for (unsigned j = 0; j < 3; ++j)
+        {
+          Rotation_matrix_at_node[j_node](3 + i, 3 + j) = b22(i, j);
+        }
+      }
+    } // end fill_in_rotation_matrix_at_node_with_basis
+
+    /// Pointer to the `parent' finite element which this is a helper force
+    FiniteElement* Parent_element_pt;
+
+    /// The number of nodes (that we store rotation data for) in the fvk element
+    /// that uses this helper
+    unsigned Nnode;
+
+    /// Vector containing boundary parametrised location for each node
+    Vector<double> Boundary_coordinate_of_node;
+
+    /// Vector containing boundary parametrisation at each node
+    Vector<C1CurviLine*> Nodal_boundary_parametrisation_pt;
+
+    /// Vector containing <rotation matrix at each node>
+    Vector<DenseMatrix<double>> Rotation_matrix_at_node;
   };
-
-  /// \short Specialisation of CurvilineGeomObject for half a circle.
-  class CurvilineCircleTop : public CurvilineGeomObject
-  {
-  public:
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleTop()
-      : CurvilineGeomObject(), Radius(1.0), Clockwise_zeta(false)
-    {
-    }
-
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleTop(const double& radius, const bool& clockwise_zeta)
-      : CurvilineGeomObject(), Radius(radius), Clockwise_zeta(clockwise_zeta)
-    {
-    }
-    /// \short Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineCircleTop(TimeStepper* time_stepper_pt)
-      : CurvilineGeomObject(time_stepper_pt)
-    {
-    }
-
-    /// Broken copy constructor
-    CurvilineCircleTop(const CurvilineCircleTop& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineCircleTop");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineCircleTop&)
-    {
-      BrokenCopy::broken_assign("CurvilineCircleTop");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineCircleTop() {}
-
-    /// \short Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = -Radius * std::sin(zeta[0]);
-      r[1] = Radius * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        r[0] *= -1;
-      }
-    }
-
-    /// \short Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius * std::cos(zeta[0]);
-      drdzeta[1] = -Radius * std::sin(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[0] *= -1;
-      }
-    }
+ 
 
 
-    /// \short 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = Radius * std::sin(zeta[0]);
-      drdzeta[1] = -Radius * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[0] *= -1;
-      }
-    }
+ 
+////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////
 
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The arc length (parametric parameter) for the upper semi circular arc
-      return (Clockwise_zeta ? atan2(x[0], x[1]) : atan2(-x[0], x[1]));
-    }
-
-  private:
-    double Radius;
-    bool Clockwise_zeta;
-  };
-
-  /// \short Specialisation of CurvilineGeomObject for half a circle.
-  class CurvilineCircleBottom : public CurvilineGeomObject
-  {
-  public:
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleBottom()
-      : CurvilineGeomObject(), Radius(1.0), Clockwise_zeta(false)
-    {
-    }
-
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleBottom(const double& radius, const bool& clockwise_zeta)
-      : CurvilineGeomObject(), Radius(radius), Clockwise_zeta(clockwise_zeta)
-    {
-    }
-
-    /// \short Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineCircleBottom(TimeStepper* time_stepper_pt)
-      : CurvilineGeomObject(time_stepper_pt)
-    {
-    }
-
-    /// Broken copy constructor
-    CurvilineCircleBottom(const CurvilineCircleBottom& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineCircleBottom");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineCircleBottom&)
-    {
-      BrokenCopy::broken_assign("CurvilineCircleBottom");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineCircleBottom() {}
-
-    /// \short Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = Radius * std::sin(zeta[0]);
-      r[1] = -Radius * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        r[1] *= -1;
-      }
-    }
-
-    /// \short Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = Radius * std::cos(zeta[0]);
-      drdzeta[1] = Radius * std::sin(zeta[0]);
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-
-    /// \short 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius * std::sin(zeta[0]);
-      drdzeta[1] = Radius * std::cos(zeta[0]);
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The arc length (parametric parameter) for the upper semi circular arc
-      return (Clockwise_zeta ? atan2(x[0], x[1]) : atan2(x[0], -x[1]));
-    }
-
-  private:
-    double Radius;
-    bool Clockwise_zeta;
-  };
-
-
-  /// \short Specialisation of CurvilineGeomObject for half a circle.
-  class CurvilineCircleRight : public CurvilineGeomObject
-  {
-  public:
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleRight()
-      : CurvilineGeomObject(), Radius(1.0), Clockwise_zeta(false)
-    {
-    }
-
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineCircleRight(const double& cx,
-                         const double& cy,
-                         const double& radius,
-                         const bool& clockwise_zeta = false)
-      : CurvilineGeomObject(),
-        Cx(cx),
-        Cy(cy),
-        Radius(radius),
-        Clockwise_zeta(clockwise_zeta)
-    {
-    }
-    /// \short Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineCircleRight(TimeStepper* time_stepper_pt)
-      : CurvilineGeomObject(time_stepper_pt)
-    {
-    }
-
-    /// Broken copy constructor
-    CurvilineCircleRight(const CurvilineCircleRight& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineCircleRight");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineCircleRight&)
-    {
-      BrokenCopy::broken_assign("CurvilineCircleRight");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineCircleRight() {}
-
-    /// \short Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = Cx + Radius * std::cos(zeta[0]);
-      r[1] = Cy + Radius * std::sin(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        r[1] *= -1;
-      }
-    }
-
-    /// \short Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius * std::sin(zeta[0]);
-      drdzeta[1] = Radius * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-
-    /// \short 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius * std::cos(zeta[0]);
-      drdzeta[1] = -Radius * std::sin(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The arc length (parametric parameter) for the right semi circular arc
-      return (Clockwise_zeta ? atan2(x[0], x[1]) : atan2(-x[0], x[1]));
-    }
-
-  private:
-    double Cx;
-    double Cy;
-    double Radius;
-    bool Clockwise_zeta;
-  };
-
-
-  /// \short Specialisation of CurvilineGeomObject for half a circle.
-  class CurvilineEllipseTop : public CurvilineGeomObject
-  {
-  public:
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineEllipseTop()
-      : CurvilineGeomObject(), Radius1(1.0), Radius2(1.0), Clockwise_zeta(false)
-    {
-    }
-
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineEllipseTop(const double& radius1,
-                        const double& radius2,
-                        const bool& clockwise_zeta = false)
-      : CurvilineGeomObject(),
-        Radius1(radius1),
-        Radius2(radius2),
-        Clockwise_zeta(clockwise_zeta)
-    {
-    }
-    /// \short Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineEllipseTop(TimeStepper* time_stepper_pt)
-      : CurvilineGeomObject(time_stepper_pt)
-    {
-    }
-
-    /// Broken copy constructor
-    CurvilineEllipseTop(const CurvilineEllipseTop& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineEllipseTop");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineEllipseTop&)
-    {
-      BrokenCopy::broken_assign("CurvilineEllipseTop");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineEllipseTop() {}
-
-    /// \short Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = -Radius1 * std::sin(zeta[0]);
-      r[1] = Radius2 * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        r[0] *= -1;
-      }
-    }
-
-    /// \short Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius1 * std::cos(zeta[0]);
-      drdzeta[1] = -Radius2 * std::sin(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[0] *= -1;
-      }
-    }
-
-
-    /// \short 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = Radius1 * std::sin(zeta[0]);
-      drdzeta[1] = -Radius2 * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        drdzeta[0] *= -1;
-      }
-    }
-
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The arc length (parametric parameter) for the upper semi circular arc
-      return (Clockwise_zeta ? atan2(x[0] / Radius1, x[1] / Radius2) :
-                               atan2(-x[0] / Radius1, x[1] / Radius2));
-    }
-
-  private:
-    double Radius1;
-    double Radius2;
-    bool Clockwise_zeta;
-  };
-
-  /// \short Specialisation of CurvilineGeomObject for half a circle.
-  class CurvilineEllipseBottom : public CurvilineGeomObject
-  {
-  public:
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineEllipseBottom()
-      : CurvilineGeomObject(), Radius1(1.0), Radius2(1.0), Clockwise_zeta(false)
-    {
-    }
-
-    /// \short Constructor: Pass dimension of geometric object (# of Eulerian
-    /// coords = # of Lagrangian coords; no time history available/needed)
-    CurvilineEllipseBottom(const double& radius1,
-                           const double& radius2,
-                           const bool& clockwise_zeta = false)
-      : CurvilineGeomObject(),
-        Radius1(radius1),
-        Radius2(radius2),
-        Clockwise_zeta(clockwise_zeta)
-    {
-    }
-
-    /// \short Constructor: pass # of Eulerian and Lagrangian coordinates
-    /// and pointer to time-stepper which is used to handle the
-    /// position at previous timesteps and allows the evaluation
-    /// of veloc/acceleration etc. in cases where the GeomData
-    /// varies with time.
-    CurvilineEllipseBottom(TimeStepper* time_stepper_pt)
-      : CurvilineGeomObject(time_stepper_pt)
-    {
-    }
-
-    /// Broken copy constructor
-    CurvilineEllipseBottom(const CurvilineEllipseBottom& dummy)
-    {
-      BrokenCopy::broken_copy("CurvilineEllipseBottom");
-    }
-
-    /// Broken assignment operator
-    void operator=(const CurvilineEllipseBottom&)
-    {
-      BrokenCopy::broken_assign("CurvilineEllipseBottom");
-    }
-
-    /// (Empty) destructor
-    virtual ~CurvilineEllipseBottom() {}
-
-    /// \short Position Vector w.r.t. to zeta:
-    virtual void position(const Vector<double>& zeta, Vector<double>& r) const
-    {
-      r[0] = Radius1 * std::sin(zeta[0]);
-      r[1] = -Radius2 * std::cos(zeta[0]);
-      // Zeta -> - Zeta
-      if (Clockwise_zeta)
-      {
-        r[1] *= -1;
-      }
-    }
-
-    /// \short Derivative of position Vector w.r.t. to zeta:
-    virtual void dposition(const Vector<double>& zeta,
-                           Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = Radius1 * std::cos(zeta[0]);
-      drdzeta[1] = Radius2 * std::sin(zeta[0]);
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-
-    /// \short 2nd derivative of position Vector w.r.t. to coordinates:
-    /// \f$ \frac{d^2R_i}{d \zeta_\alpha d \zeta_\beta}\f$ =
-    /// ddrdzeta(alpha,beta,i).
-    /// Evaluated at current time.
-    virtual void d2position(const Vector<double>& zeta,
-                            Vector<double>& drdzeta) const
-    {
-      drdzeta[0] = -Radius1 * std::sin(zeta[0]);
-      drdzeta[1] = Radius2 * std::cos(zeta[0]);
-      if (Clockwise_zeta)
-      {
-        drdzeta[1] *= -1;
-      }
-    }
-
-    /// Get s from x for part 0 of the boundary (inverse mapping - for
-    /// convenience)
-    double get_zeta(const Vector<double>& x) const
-    {
-      // The arc length (parametric parameter) for the upper semi circular arc
-      return (Clockwise_zeta ? atan2(x[0] / Radius1, x[1] / Radius2) :
-                               atan2(x[0] / Radius1, -x[1] / Radius2));
-    }
-
-  private:
-    double Radius1;
-    double Radius2;
-    bool Clockwise_zeta;
-  };
 } // namespace oomph
 
 #endif
